simplify and remove need for store
