+++ src/analytics/operations.ts
